requirements.txt
fastapi==0.109.2
Flask==3.0.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
keras==2.15.0
matplotlib==3.8.2
ml-dtypes==0.2.0
numpy==1.26.4
packaging==23.2
python-multipart==0.0.9
pillow==10.2.0
requests==2.31.0
setuptools==68.2.0
tensorboard==2.15.2
tensorflow==2.15.0
uvicorn==0.27.1
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.14.1
